Moved sonification to Extensions directory.
