+++ js/Extensions/Sonification/InstrumentDefinitions.js
@@ -11,7 +11,7 @@
  * */
 'use strict';
 import Instrument from './Instrument.js';
-import utilities from './utilities.js';
+import utilities from './Utilities.js';
 var instruments = {};
 ['sine', 'square', 'triangle', 'sawtooth'].forEach(function (waveform) {
     // Add basic instruments
