+++ ts/Extensions/Sonification/Options.ts
